flask-api/requirements.txt
# flask-api/requirements.txt

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0

# MySQL driver
mysqlclient==2.2.1

# Alembic for migrations
alembic==1.13.1

# Redis client
redis==5.0.1

# ChromaDB client
chromadb==0.4.22

# HTTP client (for OpenRouter calls)
requests==2.31.0

# Password hashing
bcrypt==4.1.2

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

